(frontend) accept/decline function
for project manager
and client

+++ application/views/dailylog/dailylog_index.php
@@ -29,9 +29,11 @@
                                         <div class="p-2">
                                             <h5 id="job_name_display"><?php echo $getJobName[$job_id] ?></h5>
                                         </div>
+                                        <?php if($this->session->userdata('role') == 1 || $this->session->userdata('role') == 2 || $this->session->userdata('role') == 3){?>
                                         <div class="ml-auto p-2">
                                             <button onclick = "window.location.href='<?php echo base_url('dailylog/dailylog_add_index/')?><?php echo $job_id?>';" class="btn-lg btn-success"> Add Daily Log</button>
                                         </div>
+                                        <?php } ?>
                                     </div>
                                 </div>
                                 <div class="card-body">
@@ -44,6 +46,9 @@
                                                     <th>Created By</th>
                                                     <th>Date Created</th>
                                                     <th>Actions</th>
+                                                    <th>Status</th>
+                                                    
+                                                    
                                                 </tr>
                                             </thead>
                                             <tbody>
@@ -55,10 +60,22 @@
                                                         <td><?php echo $getUserName[$dl['user_id']] ?></td>
                                                         <td><?php echo $dl['logdate'] ?></td>
                                                         <td>
+                                                        <?php if($this->session->userdata('role') == 1 || $this->session->userdata('role') == 2 || $this->session->userdata('role') == 3){?>
                                                             <button id="<?php echo $dl['dailylog_id'] ?>" title="Edit Details" onclick="window.location.href = '<?php echo site_url('dailylog/attachment/') ?><?php echo $job_id?>/<?php echo $dl['dailylog_id']?>'" type="button" class="btn btn-sm btn-warning"><i class="fa fa-pencil"></i> Edit Details</button>
+                                                            <?php } ?>
+                                                            
                                                             <button id="<?php echo $dl['dailylog_id'] ?>" title="View Details" onclick="window.location.href = '<?php echo site_url('dailylog/view/') ?><?php echo $job_id?>/<?php echo $dl['dailylog_id']?>'" type="button" class="btn btn-sm btn-info"><i class="fa fa-eye"></i> View Details</button>
+                                                            <?php if($this->session->userdata('role') == 1){?>
                                                             <button title="Delete" onclick="deleteDailyLog(<?php echo $dl['dailylog_id'] ?>, <?php echo $dl['job_id'] ?>)" type="button" class="btn btn-sm btn-danger"><i class="fa fa-trash"></i> Delete</button>
+                                                            <?php } ?>
                                                         </td>
+                                                        <td>
+                                                            
+                                                        <h4> <span class="badge badge-success">Accepted</span></h4>
+                                                  <h4> <span class="badge badge-warning">Pending</span></h4>
+                                                  <h4> <span class="badge badge-danger">Decline</span></h4>
+                                          </td>
+                                                            </td>
                                                     </tr>
                                                 <?php $count++;
                                                 } ?>
